Fixes #150176: Extract code related to encoded token attributes to a separate file (#150275)
