+++ src/vs/editor/common/languages/nullTokenize.ts
@@ -3,7 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { Token, TokenizationResult, EncodedTokenizationResult, ColorId, FontStyle, IState, LanguageId, MetadataConsts, StandardTokenType } from 'vs/editor/common/languages';
+import { Token, TokenizationResult, EncodedTokenizationResult, IState } from 'vs/editor/common/languages';
+import { LanguageId, FontStyle, ColorId, StandardTokenType, MetadataConsts } from 'vs/editor/common/encodedTokenAttributes';
 
 export const NullState: IState = new class implements IState {
 	public clone(): IState {
