+++ src/vs/editor/standalone/browser/standaloneThemeService.ts
@@ -7,7 +7,8 @@ import * as dom from 'vs/base/browser/dom';
 import { addMatchMediaChangeListener } from 'vs/base/browser/browser';
 import { Color } from 'vs/base/common/color';
 import { Emitter } from 'vs/base/common/event';
-import { FontStyle, TokenizationRegistry, TokenMetadata } from 'vs/editor/common/languages';
+import { TokenizationRegistry } from 'vs/editor/common/languages';
+import { FontStyle, TokenMetadata } from 'vs/editor/common/encodedTokenAttributes';
 import { ITokenThemeRule, TokenTheme, generateTokensCSSForColorMap } from 'vs/editor/common/languages/supports/tokenization';
 import { BuiltinTheme, IStandaloneTheme, IStandaloneThemeData, IStandaloneThemeService } from 'vs/editor/standalone/common/standaloneTheme';
 import { hc_black, hc_light, vs, vs_dark } from 'vs/editor/standalone/common/themes';
