+++ src/vs/editor/standalone/common/monarch/monarchLexer.ts
@@ -15,6 +15,7 @@ import { TokenTheme } from 'vs/editor/common/languages/supports/tokenization';
 import { ILanguageService } from 'vs/editor/common/languages/language';
 import * as monarchCommon from 'vs/editor/standalone/common/monarch/monarchCommon';
 import { IStandaloneThemeService } from 'vs/editor/standalone/common/standaloneTheme';
+import { LanguageId } from 'vs/editor/common/encodedTokenAttributes';
 
 const CACHE_STACK_DEPTH = 5;
 
@@ -296,7 +297,7 @@ class MonarchModernTokensCollector implements IMonarchTokensCollector {
 	private readonly _theme: TokenTheme;
 	private _prependTokens: Uint32Array | null;
 	private _tokens: number[];
-	private _currentLanguageId: languages.LanguageId;
+	private _currentLanguageId: LanguageId;
 	private _lastTokenMetadata: number;
 
 	constructor(languageService: ILanguageService, theme: TokenTheme) {
@@ -304,7 +305,7 @@ class MonarchModernTokensCollector implements IMonarchTokensCollector {
 		this._theme = theme;
 		this._prependTokens = null;
 		this._tokens = [];
-		this._currentLanguageId = languages.LanguageId.Null;
+		this._currentLanguageId = LanguageId.Null;
 		this._lastTokenMetadata = 0;
 	}
