+++ src/vs/editor/standalone/test/browser/standaloneLanguages.test.ts
@@ -7,7 +7,8 @@ import * as assert from 'assert';
 import { Color } from 'vs/base/common/color';
 import { Emitter } from 'vs/base/common/event';
 import { DisposableStore } from 'vs/base/common/lifecycle';
-import { Token, IState, LanguageId, MetadataConsts } from 'vs/editor/common/languages';
+import { Token, IState } from 'vs/editor/common/languages';
+import { LanguageId, MetadataConsts } from 'vs/editor/common/encodedTokenAttributes';
 import { TokenTheme } from 'vs/editor/common/languages/supports/tokenization';
 import { LanguageService } from 'vs/editor/common/services/languageService';
 import { ILineTokens, IToken, TokenizationSupportAdapter, TokensProvider } from 'vs/editor/standalone/browser/standaloneLanguages';
