+++ src/vs/editor/test/common/core/lineTokens.test.ts
@@ -5,7 +5,7 @@
 
 import * as assert from 'assert';
 import { IViewLineTokens, LineTokens } from 'vs/editor/common/tokens/lineTokens';
-import { MetadataConsts } from 'vs/editor/common/languages';
+import { MetadataConsts } from 'vs/editor/common/encodedTokenAttributes';
 import { LanguageIdCodec } from 'vs/editor/common/services/languagesRegistry';
 
 suite('LineTokens', () => {
