+++ src/vs/editor/test/common/modes/supports/characterPair.test.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as assert from 'assert';
-import { StandardTokenType } from 'vs/editor/common/languages';
+import { StandardTokenType } from 'vs/editor/common/encodedTokenAttributes';
 import { CharacterPairSupport } from 'vs/editor/common/languages/supports/characterPair';
 import { TokenText, createFakeScopedLineTokens } from 'vs/editor/test/common/modesTestUtils';
 import { StandardAutoClosingPairConditional } from 'vs/editor/common/languages/languageConfiguration';
