+++ src/vs/editor/test/common/modes/supports/tokenization.test.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as assert from 'assert';
-import { FontStyle } from 'vs/editor/common/languages';
+import { FontStyle } from 'vs/editor/common/encodedTokenAttributes';
 import { ColorMap, ExternalThemeTrieElement, ParsedTokenThemeRule, ThemeTrieElementRule, TokenTheme, parseTokenTheme, strcmp } from 'vs/editor/common/languages/supports/tokenization';
 
 suite('Token theme matching', () => {
