+++ src/vs/workbench/api/browser/mainThreadLanguages.ts
@@ -10,7 +10,7 @@ import { MainThreadLanguagesShape, MainContext, ExtHostContext, ExtHostLanguages
 import { extHostNamedCustomer, IExtHostContext } from 'vs/workbench/services/extensions/common/extHostCustomers';
 import { IPosition } from 'vs/editor/common/core/position';
 import { IRange, Range } from 'vs/editor/common/core/range';
-import { StandardTokenType } from 'vs/editor/common/languages';
+import { StandardTokenType } from 'vs/editor/common/encodedTokenAttributes';
 import { ITextModelService } from 'vs/editor/common/services/resolverService';
 import { ILanguageStatus, ILanguageStatusService } from 'vs/workbench/services/languageStatus/common/languageStatusService';
 import { DisposableStore, IDisposable } from 'vs/base/common/lifecycle';
